update dto and repo

+++ src/main/java/com/example/employeemanagementsystem/repository/DepartmentRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DepartmentRepo extends JpaRepository<Department,Integer> {
+public interface DepartmentRepo extends JpaRepository<Department,Long> {
